Defined ITagDomain and added Domain property to ITagType Implemented TagDomain with String, StringArray, PositiveInteger, Date and Millisecond types Changed SQLiteTagRepository and FirebirdTagRepository to use ITagDomain GetValue() method Implemented TagFactory

+++ Gnosis/src/Gnosis.Core/ITagTypeFactory.cs
@@ -7,7 +7,7 @@ namespace Gnosis.Core
 {
     public interface ITagTypeFactory
     {
-        ITagType Create(long id);
+        ITagType Create(int id);
 
         void Add(ITagType type);
         void Remove(ITagType type);
